Criar a lógica do repositório de CoreData
