+++ SoundBook/Core Data/SoundRepository.swift
@@ -1,0 +1,94 @@
+//
+//  SoundRepository.swift
+//  SoundBook
+//
+//  Created by Luca Hummel on 24/09/21.
+//
+
+import Foundation
+import CoreData
+
+class SoundRepository {
+    static let shared: SoundRepository = SoundRepository()
+    
+    private init() {}
+    
+    // MARK: - Core Data stack
+
+    lazy var persistentContainer: NSPersistentContainer = {
+        /*
+         The persistent container for the application. This implementation
+         creates and returns a container, having loaded the store for the
+         application to it. This property is optional since there are legitimate
+         error conditions that could cause the creation of the store to fail.
+        */
+        let container = NSPersistentContainer(name: "SoundBook")
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                // Replace this implementation with code to handle the error appropriately.
+                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
+                 
+                /*
+                 Typical reasons for an error here include:
+                 * The parent directory does not exist, cannot be created, or disallows writing.
+                 * The persistent store is not accessible, due to permissions or data protection when the device is locked.
+                 * The device is out of space.
+                 * The store could not be migrated to the current model version.
+                 Check the error message to determine what the actual problem was.
+                 */
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+
+    // MARK: - Core Data Saving support
+
+    func saveContext () {
+        let context = persistentContainer.viewContext
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                // Replace this implementation with code to handle the error appropriately.
+                // fatalError() causes the application to generate a crash log and terminate. You should not use this function in a shipping application, although it may be useful during development.
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+    
+    // MARK: - Core Data Getting support
+    
+    func getAllObjects() -> [Objeto] {
+        let fr = NSFetchRequest<Objeto>(entityName: "Objeto")
+        do {
+            return try self.persistentContainer.viewContext.fetch(fr)
+        } catch {
+            print(error)
+        }
+        return []
+    }
+    
+    // MARK: - Core Data Creating support
+    
+    func createObject(nome: String, intensidade: Float, imageName: String, horarioUso: String, classificacao: String) -> Objeto {
+        let objeto = Objeto(context: self.persistentContainer.viewContext)
+        
+        objeto.nome = nome
+        objeto.intensidade = intensidade
+        objeto.imageName = imageName
+        objeto.horarioUso = horarioUso
+        objeto.classificacao = classificacao
+        
+        self.saveContext()
+        return objeto
+    }
+    
+    // MARK: - Core Data deleting support
+    
+    func deleteObject(object: Objeto) {
+        self.persistentContainer.viewContext.delete(object)
+        self.saveContext()
+    }
+}
