+++ SoundBook/View Controllers/HomeScreenViewController.swift
@@ -30,6 +30,7 @@ class HomeScreenViewController: UIViewController, UISearchBarDelegate, UITableVi
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         tableView.dataSource = self
         tableView.delegate = self
         tableView.backgroundColor = .clear
